Module archiver::config

Details pertaining to parsing the configuration file, as well as constructing the internal objects specified by the configuration.

Structs

ArchiverConfig

Config

DropboxConfig

EMPTY_FLYSIGHTS

EMPTY_GOPROS

EMPTY_MASS_STORAGES

FlysightConfig

GoproConfig

MassStorageConfig

PushoverConfig

SendgridConfig

Constants

_IMPL_DESERIALIZE_FOR_ArchiverConfig

_IMPL_DESERIALIZE_FOR_Config

_IMPL_DESERIALIZE_FOR_DropboxConfig

_IMPL_DESERIALIZE_FOR_FlysightConfig

_IMPL_DESERIALIZE_FOR_GoproConfig

_IMPL_DESERIALIZE_FOR_MassStorageConfig

_IMPL_DESERIALIZE_FOR_PushoverConfig

_IMPL_DESERIALIZE_FOR_SendgridConfig

_IMPL_SERIALIZE_FOR_ArchiverConfig

_IMPL_SERIALIZE_FOR_Config

_IMPL_SERIALIZE_FOR_DropboxConfig

_IMPL_SERIALIZE_FOR_FlysightConfig

_IMPL_SERIALIZE_FOR_GoproConfig

_IMPL_SERIALIZE_FOR_MassStorageConfig

_IMPL_SERIALIZE_FOR_PushoverConfig

_IMPL_SERIALIZE_FOR_SendgridConfig

Statics

EMPTY_FLYSIGHTS

EMPTY_GOPROS

EMPTY_MASS_STORAGES