Struct archiver::ctx::Ctx

pub struct Ctx {
    pub usb_ctx: Context,
    pub cfg: Config,
    pub staging: PathBuf,
    pub notifier: Option<PushoverNotifier>,
    pub mailer: Option<SendgridMailer>,
}

Ctx is the global context object. Constructed by consuming a config::Config.

Fields
usb_ctx: Context

a USB context, used for finding and interacting with PTP devices

cfg: Config staging: PathBuf

The directory that will be used for staging files before they're uploaded.

This directory will be treated as durable! Do not set it to /tmp if you care about your files.

notifier: Option<PushoverNotifier>

An optional notifier to call on changes to uploads.

mailer: Option<SendgridMailer>

An optional mailer that will be used to send reports when uploads finish or fail.

Methods

impl Ctx

pub fn create(cfg: Config) -> Result<Ctx, Error>

Create a new context object.

This method has many side effects, creating a libusb context, creating the staging direectory if it does not exist, etc.