Module archiver::device

Some helpers to abstract over the various types of devices that we can interact with. Much of this will probably go away at some point, and Device will become a trait instead of the enum that it is today.

This module also contains the logic for simply enumerating all currently attached devices as part of generating a plan for an upload run.

Structs

DeviceDescription

Enums

Device

Functions

attached_devices

locate_flysights

locate_gopros

locate_mass_storages

Existentials