Struct archiver::dropbox_content_hasher::DropboxContentHasher

pub struct DropboxContentHasher {
    overall_hasher: Sha256,
    block_hasher: Sha256,
    block_pos: usize,
}

Computes a hash using the same algorithm that the Dropbox API uses for the the "content_hash" metadata field.

Implements the digest::Digest trait, whose result() function returns a raw binary representation of the hash. The "content_hash" field in the Dropbox API is a hexadecimal-encoded version of this value.

Example:

extern crate digest;
extern crate archiver;
use archiver::dropbox_content_hasher::DropboxContentHasher;
use std::io::Read;
use digest::{Digest,Input,FixedOutput};

let mut hasher = DropboxContentHasher::new();
let mut buf: [u8; 4096] = [0; 4096];
let mut f = std::fs::File::open("src/lib.rs").unwrap();
loop {
    let len = f.read(&mut buf).unwrap();
    if len == 0 { break; }
    hasher.input(&buf[..len])
}
drop(f);

let hex_hash = format!("{:x}", hasher.result());
println!("{}", hex_hash);

Fields
overall_hasher: Sha256 block_hasher: Sha256 block_pos: usize
Methods

impl DropboxContentHasher

pub fn new() -> Self